% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixedexp_distribution.R
\name{dmixedexp}
\alias{dmixedexp}
\title{Returns the density of the mixed exponential distribution specified with
the given weights and means.}
\usage{
dmixedexp(x, weights, means, log = F)
}
\arguments{
\item{x}{Vector of quantiles}

\item{weights}{The weights attached to each of the individual exponential distributions}

\item{means}{The means of the individual exponential distributions}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
The density of the mixed exponential distribution
}
\description{
Returns the density of the mixed exponential distribution specified with
the given weights and means.
}
\examples{
dmixedexp(c(10,20), c(0.3,0.6,0.1), c(5,8,10))
}

