% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_exchange_rates}
\alias{get_exchange_rates}
\title{Returns exchange rates for valid currency codes. The values returned are the value of 1 unit of the
currency in the target currency. eg 1 GBP = X USD}
\usage{
get_exchange_rates(currency_codes, to_currency_code = "GBP",
  as_at_date = Sys.Date())
}
\arguments{
\item{currency_codes}{A character vector of ISO currency codes}

\item{to_currency_code}{The currency for which exchange rates are required. Default is GBL}

\item{as_at_date}{The date at which the exchange rate was valid}
}
\value{
A numeric vector of exchange rates
}
\description{
It is vectorised over currency_codes
}
\examples{
get_exchange_rates(c("USD", "HKD"), "EUR", "as.Date(2005-04-01)")
}

