% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_spec_lines_model.R
\name{get_funky_sequel_matches}
\alias{get_funky_sequel_matches}
\title{Returns a list of layers from accounts where the Sequel ID doesn't appear to be correct}
\usage{
get_funky_sequel_matches()
}
\arguments{
\item{quote_id}{The quote ID}
}
\value{
A dataframe with details of layers with strange Sequel IDs
}
\description{
Returns a list of layers from accounts where the Sequel ID doesn't appear to be correct
}
\details{
The following methods are used to check Sequel ID integrity:
- Sequel ID pattern matching
- Ensure Sequel ID year = Inception year
}
\examples{
get_funky_sequel_matches()
}

