% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_info_team.R
\name{get_info_team_policy_info}
\alias{get_info_team_policy_info}
\title{Returns the renewal information for the selected inception and expiry date ranges}
\usage{
get_info_team_policy_info(yoa_from, yoa_to, reporting_level = "class",
  reporting_values, syndicate = c(1192, 435))
}
\arguments{
\item{reporting_level}{The reporting level to query at - options are 'class', 'minorclass' or 'aatg'}

\item{reporting_values}{Vector of reporting level values}

\item{syndicate}{Vector containing syndicate IDs}

\item{inception_from}{First inception date}

\item{inception_to}{Latest inception date}
}
\value{
A dataset with risk information
}
\description{
Returns the renewal information for the selected inception and expiry date ranges
}
\examples{
get_info_team_policy_info(2013,2015,"class", "ca")
}

