% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_info_team.R
\name{get_info_team_query_from_meta}
\alias{get_info_team_query_from_meta}
\title{Returns the latest query version for the specified query ID}
\usage{
get_info_team_query_from_meta(server = "FARADAYWAREHOUSE",
  database = "FaradayInfoTeam", user = "", password = "", query_id)
}
\arguments{
\item{server}{The server to query}

\item{database}{The database}

\item{user}{The user ID (leave blank if using Windows authentication)}

\item{password}{The password (leave blank if using Windows authentication)}

\item{query_id}{The query ID}
}
\value{
A list with the following items:
- QueryID
- QueryVersion
- Query
- update_user
- update_date
}
\description{
Returns the latest query version for the specified query ID
}
\examples{
get_info_team_query_from_meta(query_id = "RenewalStatus")
}

