% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_info_team.R
\name{get_info_team_renewal_status}
\alias{get_info_team_renewal_status}
\title{Returns the renewal information for the selected inception and expiry date ranges}
\usage{
get_info_team_renewal_status(inception_from, inception_to,
  reporting_level = "team", reporting_values, syndicate = c(1192, 435))
}
\arguments{
\item{inception_from}{First inception date}

\item{inception_to}{Latest inception date}

\item{reporting_level}{The reporting level to query at - with 'team' or 'atg'}

\item{reporting_values}{Vector of reporting level values}

\item{syndicate}{Vector containing syndicate IDs}

\item{server}{The server to query}

\item{database}{The database}

\item{user}{The user ID (leave blank if using Windows authentication)}

\item{password}{The password (leave blank if using Windows authentication)}
}
\value{
A dataset with risk information
}
\description{
Returns the renewal information for the selected inception and expiry date ranges
}
\examples{
get_info_team_renewal_status(inception_from = "2015-07-01",
   inception_to = "2015-08-01",
   reporting_level = "team",
   reporting_values = "ca")
}

