% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_info_team.R
\name{get_live_policies_for_team}
\alias{get_live_policies_for_team}
\title{Returns the live policies for the specified teams}
\usage{
get_live_policies_for_team(team)
}
\arguments{
\item{team}{Vector containing the teams for which live policies are required}

\item{ccy_code}{The currency in which to show premium}

\item{as_at_date}{The exchange rate date (Exchange rates used will be end of month)}
}
\value{
A dataset with live policy details
}
\description{
Returns the live policies for the specified teams
}
\details{
The function is vectorised over team. The following columns are returned:
- Major Class
- Minor Class
- Sequel ID
}
\examples{
get_get_live_policies_for_team(c("ca","av"))
}

