% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_spec_lines_model.R
\name{get_model_url}
\alias{get_model_url}
\title{Returns the url for the quote id from the model}
\usage{
get_model_url(quote_id)
}
\arguments{
\item{quote_id}{The quote ID}
}
\value{
A string containing the url to the model
}
\description{
Returns the url for the quote id from the model
}
\details{
The function is vectorised over quote_id
}
\examples{
get_model_url(c(345,762))
}

