% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_spec_lines_model.R
\name{get_priced_layers}
\alias{get_priced_layers}
\title{Returns the priced layers}
\usage{
get_priced_layers(inception_from, inception_to)
}
\arguments{
\item{quote_id}{The quote ID}
}
\value{
A string containing the url to the model
}
\description{
Returns the priced layers
}
\details{
The function is vectorised over quote_id
}
\examples{
get_model_url(c(345,762))
}

