% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_valid_curency_codes}
\alias{get_valid_curency_codes}
\title{Returns all valid currency codes as stored within the Faraday warehouse}
\usage{
get_valid_curency_codes()
}
\description{
Returns all valid currency codes as stored within the Faraday warehouse
}
\examples{
get_valid_curency_codes()
}

