% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixedexp_distribution.R
\name{levmixedexp}
\alias{levmixedexp}
\title{Returns the mean of the mixed exponential distribution specified with
the given weights and means, with values limited to limit.}
\usage{
levmixedexp(limit, weights, means)
}
\arguments{
\item{limit}{Vector of limits}

\item{weights}{The weights attached to each of the individual exponential distributions}

\item{means}{The means of the individual exponential distributions}
}
\description{
Returns the mean of the mixed exponential distribution specified with
the given weights and means, with values limited to limit.
}
\examples{
levmixedexp(c(19,22), c(0.3, 0.5, 0.3), c(10, 4, 2))
}

