% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nil_claims.R
\name{nil_claim_survival_probabilities}
\alias{nil_claim_survival_probabilities}
\title{Calculate the following probabilities for a nil claim at a certain development period:
- the probability that it will be nil at the next development
- the probability that it will be nil at the maximum development period}
\usage{
nil_claim_survival_probabilities(claim.data, ref, dev_name, value_name,
  group = NULL)
}
\description{
Calculate the following probabilities for a nil claim at a certain development period:
- the probability that it will be nil at the next development
- the probability that it will be nil at the maximum development period
}

