% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nil_claims.R
\name{NilClaimsSurvivalProbabilities}
\alias{NilClaimsSurvivalProbabilities}
\title{Calculate the following probabilities for a nil claim at a certain development period:
- the probability that it will be nil at the next development
- the probability that it will be nil at the maximum development period}
\usage{
NilClaimsSurvivalProbabilities(claim.data, ref, dev, value, group = NULL)
}
\description{
Calculate the following probabilities for a nil claim at a certain development period:
- the probability that it will be nil at the next development
- the probability that it will be nil at the maximum development period
}

