% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixedexp_distribution.R
\name{pmixedexp}
\alias{pmixedexp}
\title{Returns the CDF value of the mixed exponential distribution specified with
the given weights and means.}
\usage{
pmixedexp(q, weights, means, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{Vector of quantiles}

\item{weights}{The weights attached to each of the individual exponential distributions}

\item{means}{The means of the individual exponential distributions}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ≤ x], otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
The CDF of the mixed exponential distribution
}
\description{
Returns the CDF value of the mixed exponential distribution specified with
the given weights and means.
}
\examples{
pmixedexp(c(10,20), c(0.3,0.6,0.1), c(5,8,10))
}

