% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixedexp_distribution.R
\name{qmixedexp}
\alias{qmixedexp}
\title{Returns the quantile function value of the mixed exponential distribution specified with
the given weights and means.}
\usage{
qmixedexp(p, weights, means, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{Vector of probabilities}

\item{weights}{The weights attached to each of the individual exponential distributions}

\item{means}{The means of the individual exponential distributions}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X ≤ x], otherwise, P[X > x].}

\item{log.p}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
The quantile function value of the mixed exponential distribution
}
\description{
Returns the quantile function value of the mixed exponential distribution specified with
the given weights and means.
}
\examples{
pmixedexp(c(0.6, 0.75, 0.9, 0.99), c(0.3, 0.6, 0.1), c(5, 8, 10))
}

