% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixedexp_distribution.R
\name{rmixedexp}
\alias{rmixedexp}
\title{Returns random samples from the mixed exponential distribution specified with
the given weights and means}
\usage{
rmixedexp(n, weights, means)
}
\arguments{
\item{n}{A character vector of ISO currency codes}

\item{weights}{The weights attached to each of the individual exponential distributions}

\item{means}{The means of the individual exponential distributions}
}
\value{
n random samples from the mixed exponential distribution
}
\description{
Returns random samples from the mixed exponential distribution specified with
the given weights and means
}
\examples{
rmixedexp(100, c(0.3,0.5,0.2), c(10,15,24))
}

