% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distribution_fitting.R
\name{fit_distribution}
\alias{fit_distribution}
\title{Function to fit distribution to data
Provides test results from fitting distributions to data}
\usage{
fit_distribution(data, fit = "gamma", sample = 1)
}
\arguments{
\item{data}{A numeric vector containing the values to fit}

\item{fit}{Names of the distributions to fit.  Distributions "beta", "cauchy", "chi-squared", "exponential", "f", "gamma", "geometric", "lognormal", "logistic", "negative-binomial", "normal", "Poisson", "t" and "weibull" are recognised, case being ignored.}

\item{sample}{The proportion of the data to be used for fitting. A random sample will be taken. Default is 1.}
}
\value{
A dataframe containing fitted parameters and test statistics
}
\description{
Function to fit distribution to data
Provides test results from fitting distributions to data
}
\examples{
fit_distribution(data, c("poisson", "negative-binomial"))
}

